pyproject.toml
[project]
name = "oxtal-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate>=1.12.0",
    "biopython==1.83",
    "biotite==1.4.0",
    "dm-tree>=0.1.8",
    "evaluate>=0.4.6",
    "fair-esm>=2.0.0",
    "fasteners>=0.20",
    "freesasa==2.2.1",
    "gemmi>=0.6.7",
    "huggingface-hub>=0.27.0",
    "hydra-core>=1.3.2",
    "icecream>=2.1.8",
    "ipdb>=0.13.13",
    "lightning>=2.6.0",
    "ml-collections>=1.1.0",
    "modelcif>=1.6",
    "multimolecule>=0.0.8",
    "ninja>=1.13.0",
    "numpy<2",
    "pandas>=2.3.3",
    "pdbeccdutils>=1.0.0",
    "pre-commit>=4.5.1",
    "prody>=2.6.1",
    "py3dmol>=2.5.3",
    "pyrootutils>=1.0.4",
    "rich>=14.2.0",
    "rootutils>=1.0.7",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "setuptools",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "torch>=2.5",
    "torchdata>=0.11.0",
    "torchinfo>=1.8.0",
    "tos>=2.8.7",
    "tqdm>=4.67.1",
    "transformers>=4.50.0",
    "wandb>=0.23.1",
    "openfold",
    "runner",
]
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
[tool.uv.workspace]
members = ["packages/openfold", "config", "runner"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
openfold = { workspace = true }
configs = { workspace = true }
runner = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oxtal"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
[tool.black]
line-length = 99
